requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
PyPDF2>=3.0.0
xlsxwriter>=3.0.0
google-genai
